requirements.txt
aioboto3==11.2.0
aiobotocore==2.5.0
aiohttp==3.8.5
aioitertools==0.11.0
aioredis==2.0.1
aiosignal==1.3.1
alembic==1.11.2
annotated-types==0.5.0
anyio==3.7.1
async-timeout==4.0.3
asyncio==3.4.3
asyncpg==0.28.0
attrs==23.1.0
bcrypt==4.0.1
boto3==1.26.76
botocore==1.29.76
certifi==2023.7.22
cfgv==3.3.1
charset-normalizer==3.2.0
click==8.1.6
coverage==7.2.7
distlib==0.3.7
dnspython==2.4.2
email-validator==2.0.0.post2
exceptiongroup==1.1.2
fastapi==0.101.0
filelock==3.12.2
frozenlist==1.4.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.17.3
httptools==0.6.0
httpx==0.24.1
identify==2.5.26
idna==3.4
iniconfig==2.0.0
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
mock==5.1.0
multidict==6.0.4
nodeenv==1.8.0
packaging==23.1
pathspec==0.11.2
platformdirs==3.10.0
pluggy==1.2.0
pre-commit==3.3.3
pydantic==2.1.1
pydantic-extra-types==2.0.0
pydantic-settings==2.0.2
pydantic_core==2.4.0
PyJWT==2.8.0
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
PyYAML==6.0.1
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
timestamp==0.0.1
tomli==2.0.1
typing_extensions==4.7.1
urllib3==1.26.9
uvicorn==0.23.2
uvloop==0.17.0
virtualenv==20.24.2
wrapt==1.15.0
yarl==1.9.2
